Processing of the 'publication_info' field has been modified
Function [449-473] convert_to_list() was implemented as a common
function, processing fields which can be dicts or lists.
This function was optimized for the processing of authors.
Currently it is removed.
Subfunction fix_dict_values() also removed.
Instead, cheking the type of 'publication_info' field has been
moved to process_journals() function.
Tests showed that in 'publication_info' subfields
all strings are corrected, so there is no need to execute fix_string
for each field.