pyDKB: change the Messages logic to decode/encode
Now message methods do the following:
* init    -- save the original data
* decode  -- try to interpret the original data as "coded" and transform it to
  the usable format (e.g. JSON string -- to Python dict object)
* encode  -- try to interpret the original message as the real message content
  and "code" it to the given format (e.g. Python dict object -- to the string)
* content -- try to return the decoded message

If the message is in the "native" format for the given message type, the
original data will be taken as "decoded".

If the message was decoded/encoded successfully, both "decoded"/"encoded"
parameters will be set: one -- to the obtained value, another -- to the
original message.

For now only one "code" type is defined -- STRING, as we transmit data from
stage to stage via stdin/stdout in the plain text format.