pyDKB: implement output logic for Processor Stage.
Major changes:
* process() method has became a @staticmethod
  ...to allow usage of output() function from within the process()
  implementation outside the class definition.
  It means that the outside definition for the function now becomes
  {{{
  def process(stage, message):
      ...
      stage.output(output_messages)
  }}}

* process() method must return True or False
  ...to indicate if the processing was successful or not. If the
  processing has failed, all the output messages are dropped.

The output messages are buffered in __output_buffer parameter, and
pushed forward only when the input message processing finished
successfully.

Output messages separator and end-of-process (EOP) marker for
(s)treaming mode are hardcoded for now (\n and \0 respectively), but
should be made confugurable later. The separator for input messages is
the same as for output messages.

In (f)ile mode output goes to files with the same basename as the
original one but with another extention (it is determined by the output
message type). The files will appear in the same directory, as the
original ones, or in the directory specified with --output-dir.

The example usage of the Stage class is demonstrated in
test/pyDKB/json2TTL.py