pyDKB: add stop() method (fix issue with the last HDFS output file)
The issue was that due to the __output generator never reaches last
hdfs.putfile() and the last file wasn`t uploaded to the HDFS.

So stop() method was added. It calls close() method of all the objects
in __stoppable list (currently -- __input and __output generators).

generator.close() throws GeneratorExit exception, so the construction
`try...yield...finally <exit_statements>` catches it and execute
the <exit_statements>.

It also catches any other exception, so the generators will be finished
correctly in case of any other exception.