pyDKB/stage: fix issue with in/out JSON files (issue #24, NDJSON).
If output multiple JSON messages to a single file, we end up with NDJSON
(due to the message-oriented architecture of the pyDKB), but input
function expected to see a true JSON file (JSON Array or a single JSON
object).

Now NDJSON is a default format for the transitional files, yet the true
JSON files still can be used for the input.