Stage 095_datasetsInfoAMI added
- stage is integrated with pyDKB library
- gets the input from stage 091 and add PhysParametersVals from AMI
  to each dataset NDJSON string

Note:
----
AMI GetPhysicsParamsForDataset function works with containers. But from
stage 093 we get datasets. To transform dataset name into container name
the _tid{...} part is removed from the full dataset name. So, we obtain
crossSection, getFiltEff, and other parameters from AMI for containers.

Suggesting that containers consist of the datasets of the same type, and
the PhysParametersVals are the same for contaier and for all its
datasets, we extand the initial parameters of datasets with
PhysParametersVals from AMI.