DF/09: reimplement `squash_records` function.
Two main reasons:
* it was not a generator, and although all the other functions allowed
  us to operate with records received from the Oracle 'on the fly', we
  still had to consume all the data for 'datasets' query at once and
  loop over them stored in memory;

* it was implemented strictly for the given use case (when we have input
  records with these three fields: 'taskid', 'type', 'datasetname').
  The current implementation also has pre-defined key fields and the
  rule to fold the original record, but it is formalized and thus can
  easily be reconfigured.

The main idea of generalization of any of the functions of this stage is
to make it possible to use it not only with given 'tasks' and 'datasets'
SQL queries, but also with any set of queries, and to allow us to
introduce any changes to the query without changing the source code of
the connector. The task is still incomplete, but we`re on the way.