DF/09: refine relative path to query files handling.
The relative path was relative to current working directory, which is
not a good starting point (as the stage is usually run not from its own
directory, but from `../run/` or `../000_kafka/` or wherever else.

Now the relative path is considered to be related to the directory
containing the config file.