pyDKB/enhance: custom readline implementation
custom_readline() allows to use custom EOM.

Non-blocking stdin was implemented in case of endless waiting for new
messages in a stream mode -- both parts, Kafka and pyDKB, will wait
until required size of buffer will be filled. So it might freeze the process.

The system responds each time when new message appears for reading.
Now each time it reads only one message and then -- continues to listen
to the stream.

The idea of leaving read(1) was rejected for the reason of inadvisable memory
usage.