DF/09: replace ConfigPareser with SafeConfigParser.
It is recommended to use SafeConfigParser instead of CongfigParser, as
it has "more predictable" implementation of "magical interpolation"
(string substitution in config files).
Though currently we do not use this possibility, it is better to use
more safe variant so that when someone decides to use it, we were
alright with this.