pyDKB/enhance: ArgumentParser epilog
I decided to use argparse.RawDescriptionHelpFormatter instead of
argparse.RawTextHelpFormatter because with the use of the second one
'help' looks a bit... uhmm... ugly. In the terminal lines that requieres
a break tries to fill the whole width of the window.

Epilog -- text to display after the argument help --  contains a table of
shortcuts. This table contains pre-defined EOP and EOM markers, source and
destination types depended from a chosen mode.