pyDKB/enhance: ArgumentParser epilog
I decided to use argparse.RawDescriptionHelpFormatter instead of
argparse.RawTextHelpFormatter because with the use of the second one
'help' looks a bit... uhmm... ugly. In the terminal lines that requieres
a break tries to fill the whole width of the window.

If the epilog content looks strange in the middle of the window, I'll fix
it and move it to the left.