pyDKB/enhance: KeyboardInterrupt handling
If an interruption -- invoked by user -- occurs and it's no handled,
Python produces standard output with traceback that looks ugly and
useless. The exception handler produces a short log message
and invokes exit.

In case of a keyboard interruption by hands the system throws
an (INFO) message. It is essential for keeping program logs -- if it's
just a simple situation when user pressed Ctrl+C in the terminal.