ES/event_summary some fixes
- 'should' clause removed from the aggregated query
We need the query to behave like A and B and C

Was:
  "bool": {
    "must": [ A, B ],
    "should": [ C ],
  }

Here "should" clause affects only the order of the query results (https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-bool-query.html).
To make its conditions required, it must be moved inside the "must" clause:
  "bool": {
    "must": [
      A, B, C
    ]
  }

- 'hashtag_list' clause moved from 'should' to 'must' section
- .keyword suffix removed from 'campaign'
(if we use 'campaign.keyword', then we have to input case-sensitive name of campaign,
and the usage of just 'campaign' allows to input case-insensitive campaign name,
that is better for searching).