pyDKB: introduce 'communication' sumbodule (InputStream class).
One day we are going to add Connector Stages, which also need to
communicate with Supervisor (same way as the Processor does).
Encapsulated this functional within Abstract Stage is not very good, as
Source connector have no input data, and Sink one does not have
*output*; while Processor has them both. So it is better to move this
functional into a separate class and use it whenever it is needed.

pyDKB.dataflow.communication subclass is aimed to encapsulate functions,
  related to the communication between Supervisor (Kafka app) and Worker
  (pyDKB Stage).

pyDKB.dataflow.messages is moved into ..communication, as Message object
  is mostly related to the communication process, not anything else.

..communication.Stream is a base class for input/output streams of
  messages.

..communication.InputStream is a implementation of input messages
  stream.

..communication.Stream() is a constructor function for input/output
  stream implementation objects, choosing appropriate implementation
  basing on the passed open file descriptor mode.