pyDKB/hdfs: replace `getfile()` with `File()`.
When we work with HDFS as a temporary storage for some transitional
data, most of the time wa want to get temporary local copy of the HDFS
file for work -- not real file. Thus `getfile()`, that used to create
local copy of the file for further work, is replaced with `File()`, that
returns `tempfile.TemporaryFile` with the content of the original HDFS
file. `TemporaryFile` does create local temporary file, but it also
takes care of removing it as soon as the file was closed.

As `ProcessorStage.__hdfs_in_files()` must no longer take care of
removing the temporary file, `try/finally` clause was removed.