Codestyle: fix W605 (invalid escape sequence).
As far as I understand, this warning appears when backslash is used in
an ordinary string as an ordinary character.

It makes sense -- as one might not remember all the escape sequences
exist (like `'\n'`, `'\\'`, `'\xYY'`, etc.), and can unintentionly put
it into an ordinary string (say, `'\def'` is an ordinary string of 4
characters, while `'\abc'` icontains 3 characters, first of which is
ASCII Bell (BEL) symbol).

Thus it is suggested to use `'\\'` in an ordinary string, when sole
backslash is needed, or use raw strings (like this: `r'\abc'`).

Note: `re.compile()` accepts standard Python escape sequences, like
`'\n'` (so `r'\n'` is treated same way as pure `'\n'`).