TravisCI: fix issue with single error being ignored.
When codestyle check produces single error message, `$num_errors`
remains zero due to lack of newline symbol in the end of error message.
`echo -n` was used to get zero value for empty line, but unfortunately
it affects non-empty line as well.

Now we simply check if there anything in the `$errors` variable.