pyDKB/*/AbstractStage: improve output_error() method.
It used to output with `logLevel.X` first and last N lines:
* ERROR: 3 lines
* DEBUG: 8 lines (but the 3 with ERROR level)
* TRACE: all the others

Now it outputs with `logLevel.X` only the last N lines:
* ERROR: 1 line
* DEBUG: all the others

I believe that the last line gives the most important information to
understand the content of the error; whole traceback (including first
lines) is required only for debug.
TRACE level is more to understand how things are going during the
program execution to locate unexpected behaviour, not to "get more
detais about the error".