pyDKB/*/FileProducer: fix issue with excessive empty out file.
As we needed to initialize OutputStream with real file descriptor, we
had to use temporary file with timestamp name untill we get real input
file metadata. To avoid it we might use in-memory objects like StringIO,
but real issue is that we need OutputStream object before we have enough
information to create a proper output file; and, in fact, OutputStream
can do good enough without real file descriptor for a while.

So instead of creating fake file descriptor, we use sub-configured
OutputStream -- it will raise exception if we try to use functionality
that requires real file descriptor, but works fine if we do not.

It also allows us to call `reconfigure()` in the `__init__()` method.