pyDKB/*/ProcessorStage: specify default output directory.
It is better to separate output files from input ones by defaut, as for
stages with same input and output data format previout default behaviour
lead to collisions: output file had same name and same path, as the
input one.

Now default behaviour is to place output files to the 'out' subdirectory
of the current directory.