pyDKB/dataflow: introduce 'communication' submodule.
One day we are going to add Connector Stages, which also need to
communicate with Supervisor (same way as the Processor does).
Encapsulating this functional within Abstract Stage is not very good, as
Source connector have no *input* data, and Sink one does not have
*output*; while Processor has them both. So it is better to move this
functional into a separate class and use it whenever it is needed.

pyDKB.dataflow.communication submodule is aimed to encapsulate functions,
  related to the communication between Supervisor (Kafka app) and Worker
  (pyDKB Stage).

pyDKB.dataflow.messages is moved into 'communication', as Message object
  is mostly related to the communication process, not anything else.