DF/69: custom EOP and EOB implementation
EOP -- End-of-process
EOB -- End-of-batch (old: $DELIMITER)

For `read -d` in bash EOB (as a delimiter)  should be not more that one byte.

The previous version of a script causes contradictions connected with
`mode dependent` behaviour. Despite the fact that the arguments for
determining custom markers are counted from the second, the script
switches to a file mode only if there are no arguments at all. In
that case we just can't set any custom marker for a file mode.