Added ATLAS+data_samples directory with data samples ontology and revised ATLAS ontology.
This version of ATLAS ontology is expanded to contain majority of all data sample metadata that is present in DEFT and Rucio, according to their schemas.
Intended use: for experiments in data integration. This ontology may replace the current ATLAS ontology if we ever discover the need to store and analyze all the metadata included in it.