pyDKB: fix logic of submodule import.
For now we have two main submodules: 'common' and 'dataflow'.
If we failed to import 'common', most likely 'dataflow' functionality is
broken.
If we failed to import 'dataflow', there`s very little can be done
without it -- thus it is better to fail immediatly.

If one day we need to use submodules independently, it is possible to:
* not to import them in library init file OR
* not to raise ImportError, leaving only warning message.