pyDKB: fix issue with empty command line argument list.
If run Stage without any parameters, it fails to `configure()` with
error:
```
  File "./../../pyDKB/dataflow/stage/ProcessorStage.py", line 179, in
configure
    self.__input = consumer.ConsumerBuilder(vars(self.ARGS)) \
TypeError: vars() argument must have __dict__ attribute
```

To fix it we need to ensure ARGS is initialized in any scenario.
We could initialize ARGS as `argparse.Namespace()` during the object
initialization, but it is better to run parse_args (just in case stage
has some required parameters).