pyDKB/logging: change format suffix handling.
Instead of determinig the suffix every now and then from the format, we
can store it in object attribute variable (`_suffix`) and reuse wherever
it is needed.

As `Formatter` does not support format change on the fly, it is safe
enough (suffix won't last longer than the format it was taken from).