pyDKB/*/FileConsumer: turn `_filenames_form_dir()` into generator.
Not a big deal, but most of FileConsumer methods, related to input
files, are generators -- so why not. Now instead of checking all files
in the directory and creating a list of those with proper extension, it
yields first file name that fits requirements, and checks next file only
when asked about it.