pyDKB/*/communication: fix empty source detection (custom EOM).
`FileConsumer.source_is_empty()`, used to check if current (file) data
source is fully read, checked current cursor position against file size;
but when custom EOM is used, `custom_readline()` reads data into local
buffer and thus file can be fully read, while not all the values from
the buffer are processed.

To fix this issue new method `InputStream.is_readable()` is added with
different implementations: for ordinar file reading (used to work with
`'\n'` or empty EOM) and for `custom_readline()` generator.

Now if current source is bound to a Stream object, it is checked via the
Stream method; if not -- directly, as it used to be. File size is not
saved in the `current_file` hash as it shouldn't be checked here that
often as when there were no such functionality in the `InputStream`.