es: rename indices (tasks_production, tasks_analysis).
Mapping for these indices is essentially the same (as we store there
same data from same sources); thus there is no need to have two
identical mapping templates, when we can use same pattern for both
indices.

Having two files with mapping is not good: changing mapping we need to
change it twice.

ES prior to 6.0 does not allow to have "OR" in the template: we can not
say "prodsys_rucio_ami OR analysis" in any way. Thus we rename indices
accordingly to: "tasks_production" and "tasks_analysis", and set index
template to "tasks_*" for the mapping.