DF/009: make SQL queries parametrized via config file.
We need two almost identical queries for production and analysis tasks;
they differ only it the condition on `pr_id` field.
To avoid duplicate files in the repository (which may lead to mistakes
like "changed one file and forget about another"), the query must be
parametrized in some way. As we already have config file for this stage,
seems reasonable to put query parameters into it.

Parametric pieces of queries must be specified with
`%(<parameter_name>)s`, and values specified in `queries.params` section
of the config file: `<parameter_name>: <parameter_value>`.`

When parameters are used, one must make sure that no precent signs in the
query will produce error and escape them, if needed ('%%' instead of '%').