pyDKB/*/consumer: fix get_{source,stream} logic.
`get_source()` tried to skip empty sources, but as the functional to
detect if source is empty was moved to the InputStream object, we got
collision: current `_stream` was not reset until `get_source()` returns
new non-empty source; but `get_source()` relied on the InputStream's
methods and always saw that current source (or, actually, current
stream) is empty.

Now logic is simplified, recursion-like method calls are removed:
* `get_source()` returns current source (or, if source is not defined
  yet, the next one);
* `get_stream()` returns stream "as is" (or, if stream is not defined
  yet, initializes it with current source and returns new one);
* `next_source()` returns next source;
* `reset_stream()` resets stream to the next source.

Yet to read new message from the InputStream object, we still need to reset
it to the "nearest readable source" (or else we can not say if it is
just current source is over or we run out of sources at all). So new
method is added:
* `get_readable_stream()` returns stream, resetted to the nearest
  non-empty source.