DF/69: custom EOP/EOB implementation
*Adding help
*Following options were implemented:

EOP -- End-of-process
EOB -- End-of-batch (old: $DELIMITER)
BATCHMODE -- an option to support batch-mode processing.

For `read -d` in bash EOB (as a delimiter)  should be not more that one byte.
For batch-mode processing that option should be enabled:

-b|--batch (e)enabled|(d)isabled
Default: disabled

If EOB (-B|--eob) is specifiend with a custom marker
(not "\x17" -- default), batch-mode will be enabled anyway.

In other cases batch-mode should be enabled.
Note: Empty EOB = disabled batch mode.