docs/pyDKB: improve docstrings (pyDKB.common.Type).
As the module consists of Type class only, "usage" was moved to the
module docstring -- after all, it explains destination of the module,
while class Type is just a realization.

NOTE: by default `__init__()` method description is missed in docs, so
when needed, constructor paremeters should be added to class docstring.