API: add implementation for `storages.es.init()` with hardcoded config.
As `storages.es.STORAGE_NAME` is used in `config` submodule, is must be
declared before it is imported, so import statement was moved.

As `elasticsearch.Elasticsearch()` can raise various exceptions (e.g.
`ValueError`, `TransportError`), to catch them all
`try/except Exception` is used. For debug means whole traceback is
logged by server, while user response includes only the error message.

ES coniguration is to be specified in `es_params` file as Shell
variables (example is added to the repository). In the future it should
be moved to a proper config file.