DF/10: $EOM/$EOP usage
About $EOP output:

While we have only 1 message (so-called JSON-array) EOM and EOP
should be just sent after that message. It's possible to write both of them
into $tmp-file, (or) but it would be better to leave only EOM in $tmp
and to output EOP explicitly (just to be prepared for a case when we'll divide
somehow "data channel" and "command channel").