DF/69: Special symbols handling
In case of handling escape sequences for custom markers
and because we should handle somehow whitespace symbols
like '\0', '\n', etc., we should choose one proper approach
and stick to it.

The most suitable way is to put escape sequences to variables and then
interpret them when marker is used.

Like this:
{{{

  EOProcess='\x06'
  <...>
  while read -d $(echo -ne "$DELIMITER") line ; do
    <...>
    echo -ne "$EOProcess"
  done

}}}

This solution also works with whitespace symbols.