docs: move content of /Docs directory to /docs/misc.
`/Docs` contained MS Office files with miscellaneous information:
instructions for team members, description of data schemes,
presentations created for working meetings.

All these files were splitted in two categories: 'instructions' (for
documents containing explanations for team members of "how to work with
this or that") and 'presentations'.

Document `ontology changelog.docx` is removed for its meaning is long
lost, but speculative it was intended to explain one specific commit.