DF/19: change default value for EOP to '\0'.
According to the default value of the EOP marker for supervised
dataflow.

`data4es` start script is changed accordingly (to filter introduced EOP
marker in the pipeline).

Looks wierd -- add marker and immediately filter it out -- but one day
we want to delegate supervisor tasks to Kafka supervisors, and then we
don`t want to miss this marker.