docs/pyDKB: make 2nd code example clearly connected to the 1st one.
It could be not obvious for reader how the "main" part of the example
application is connected to the "functional" part, provided in another
code snippet. To make things more clear, a nod to  "functional" part is
included to the second sample. Introduction to the code sample also
extended, and "new" lines are emphisized.

Including there the whole code of `my_function()` does not look a better
decision:
* if one day we decide to rewrite it, we can probably forget to do it in
  both places;
* when read the guide, two identical pieces of code usually make me
  suspicious ;) and I have to double-check all the occurences of the
  code.