pyDKB/stage: clarify help message ("optional" values).
Most of the parameter values due to the `nargs='?'` were displayed in
help message as optional, while being required.

It means, that this:
```
./stage.py -m
```

must raise error, but help said:
```
usage: json2TTL.py [-h] [-m [MODE]] ...
```

, as if `MODE` was optional.

It happened due to the fact that `nargs` does not have value, that would
be interpreted as "get exactly one value" AND would not return this
value as a list (as `nargs=1` would do).

Careful reading showed that "exactly one" is the default behaviour, so
omitting `nargs` gives the desired result.