DF/009/query: replase list of fields with t.* in SELECT clause.
It reduces error probability, related to necessity to add new field to a
number of places in the query. Now new fields should be added only to
the firs SELECT statement, and will be propagated to the output
automatically.

This change revealed that field 'step_id' was missed in these lists and
thus never reached stage output, though was selected in the very first
stage of SELECT.

Missed field was also missed in ES mapping; in case it is needed,
mapping was updated too.

Other changes in the output data sample are related to the hashtags
order in 'hashtag_list'.