Swapped the order of two functions in the code.
In the past, analyze_page() was used by debug/development
functions in manager(see previous commit). It will be reworked
to serve as a part of get_tables_from_text() so it is better to
position it earlier in the code. This swap is performed in a
separate commit to improve readability of the changes.