pyDKB/setup: add version info to setup util.
To make sure that `__init__` and `pyDKB-setup` share version info, it
should be taken from the same place. Unfortunately (and quite
surprisingly) `setuptools` does not have an instrument to extract this
information from `__init__` file.

Here are some variants suggested:
  https://packaging.python.org/guides/single-sourcing-package-version/

I have decided to use the one with standalone non-Python file with
version info. Better way might be the one using repo tags, but in this
case we would need a separate repo for the `pyDKB`.