pyDKB/stage: fix (m)ap-reduce mode description.
(M)ap-reduce works as follows:
 * processing script is copied to Hadoop nodes;
 * data to be processed are passed to STDIN line by line.

Data can be taken from HDFS files by Map-Reduce mechanism, and then data
themselve are passed to STDIN line by line (and thus in data files
messages should be separated by newline symbol).
This scenario is applied when --source is set to (s)tream.

If what come to STDIN are names of HDFS files, then:
 * filenames are separated by newline symbol;
 * messages in data files can be separated with any custom EOM marker.
This scenario is applied when --source is set to (h)dfs.