pyDKB/AbstractStage: move flag option check into a separate method.
The check is not obvious, so encapsulating it into a separate method
seem to be a way to make it "self-documented".

What is checked is if the argument supposed to take its value from the
command line, or use some constant value (True/Falsei or <const>, passed
to the add_argument() method). In first case it is an honest parameter,
while in second it is used as flag.