API: change directory layout.
SELinux wants to guarantee that HTTP daemons have no access to valuable
data on the server; thus it restricts access for these processes to
anything except locations known in advance.

Pre-defined directory, for example, is `/var/www` for all "content"
files and `/var/www/cgi-bin` for executables.

So it was decided to change the application layout as follows:
 * $WWW_HOME/cgi-bin/dkb.fcgi
 * $WWW_HOME/lib/dkb/api/
...in the belief that server administrator set proper home directory for
the WWW user.

It makes SELinux to allow HTTP daemon to execute `dkb.fcgi` and read
`api/`; that's what we need.