es/query: make hashtag-steps statistics accept list of hashtags.
All tasks, matching at least one of the listed hashtags, makes
contributiin into the query results, as if there were "OR" between
hashtags.

Output data sample is given for two hashtags: "gammajets" and "mc16c_cp".
Maybe not a real-life example (it would make more sense for "AND" join
operation, not for "OR"), still it is an example.

Previously used sample hashtag ("newfastcalosim") never comes with other
hashtags, so (even though it is "OR") it is changed.