Improved readability in testing hdfs overrides.
In current version the line of code is pointless because arg
is either 'source' or 'dest' which are redefined in the next few
lines. However, the code reader may not know that. Furthermore,
this line may be of use if it would be necessary to test
some other arguments' interaction with hdfs.