DF/readme: get rid of tabs in favor of spaces.
In the README there was a mix of tabs and spaces, and due to this one
using cmdline diff could never say if the alignment was broken or just
tabs look differs from that of spaces.
With tabs it is easier to align columns, yet spaces are more predictable
(for me, at lest ;) ), so tabs were replaced with spaces, not vice
versa.