DF/025: add aggregated by jobs metadata extraction (total hs06).
The HS06 we did take from Chicago ES before is actually an estimated
value for HS06*1sec per event.
But for already finished tasks it is possible to get the real value
(by aggregating task jobs metadata).

Additional metedata after this stage:
* totHS06_failed   -- CPU usage of all failed jobs;
* totHS06_finished -- CPU usage of all failed jobs;
* totHS06          -- 'finished' + 'failed'.

We could make it only 'totHS06' without distribution by status, but I
believe that both "what was really spent on this task" and "what would
be spent on this task if everything went perfect" can be useful metrics.

In Chicago ES there are individual index 'jobs_archive_YYYY-MM-DD' for
every day and it is quite a lot. To avoid hittin them all, we use task
start/end time to select indices we might need (since all the jobs
should be started and finished within this interval).

ES mapping is updated accordingly. This time I tried a new idea: add
field description to the mapping, so that one reading it could
understand what this field is and where it comes from.