DF/025: add gracefull handling of index-not-found exception.
There may be an index missed in the ES (e.g. for old tasks); retrying
will make no difference unless the index is removed from the list of
indices.
Once the index is removed, it is a different query that may
have different issues (including one more missed index), that's why the
retry counted is not decreased on this retry.