DF/025: improve index-no-found error handling (wildcard index).
If an index 'YYYY-MM-DD' is missed, it is very likely that next and
next after next and other with close dates are missed too. To avoid
removing them from the list one after one, the '-DD' part can be
replaced with wildcard '-*' -- it will include at max 30 indices, but
can save from 30 failed queries.

Also, the wildcard index does not produce 'index-not-found' error even
if no matches found. We could, of course, use 'YYYY-MM-DD*' for all
index names, it seems not to affect the query efficiency; but it is a
bit ugly, for my liking.