DF/009: refactoring (remove global variable).
'mode' was a global variable initially as it was used in different
functions. Currently it was used only in one function, still it remained
global to avoid passing it as another parameter of the function.

Now as it is added to the 'config' hash, there's no need to keep it as a
global variable any longer.