DF/009: move 'mode' from command line to config file parameters.
There's no reason to treat this parameter in any special way; besides,
in processing stages, using pyDKB "stage" abstraction, "mode" parameter
is reserved for (s)tream|(f)ile|(m)ap-reduce modes, which may be
confusing.

More reasonable would be to make 'PLAIN' policy the default one, but to
ensure there will nothing break after this change due to the lack of
'mode' parameter in this or that config file, setting default to
'SQUASH' looks as a safer variant.