pyDKB/test: add section "What to check".
When any incorrect behaviour is detected, in this section must appear a
"test case", which helps to reproduce it, before (or together) with the
patch that fixes this bug. Till the bug is not fixed, the item should be
marked with '(-)', unce fixed -- with '(+)'.

It is needed to have a list of "what was once made to work properly", so
that after any other fix one could check if he broke something or not.

Now added test case "without parameters"; expected behaviour is to
produce an error "no input files specified", but it produces:
```
$ ./json2TTL.py
(WARN) pyDKB.dataflow.cds failed (No module named
invenio_client.contrib)
Traceback (most recent call last):
  File "./json2TTL.py", line 75, in <module>
    main(sys.argv[1:])
  File "./json2TTL.py", line 56, in main
    stage.configure(args)
  File "./../../pyDKB/dataflow/stage/ProcessorStage.py", line 185, in
configure
    self.__input = consumer.ConsumerBuilder(vars(self.ARGS)) \
TypeError: vars() argument must have __dict__ attribute
```