DF/025: fix issue with TooLongFrame exception in ES requests.
If interval between start and end time is too long, there will be too
many indices in the request, which leads to the TooLongFrame exception:

  TransportError(400, u'too_long_frame_exception', u'An HTTP line is
  larger than 4096 bytes.')

For example: task 6793089.

To reduce number of indices, instead of YYYY-MM-DD* suffix we can use
YYYY-MM* (or even YYYY*, if necessary; but it was not implemented yet).