Remove a config example's duplicate.
Currently there are two examples of the config file for the
stage - the one in Utils/Dataflow/config/ is more up to date, so
the one in stage 009's readme is replaced with a link to the
former.

Noticeable difference between the two examples is the description
of the change introduced in bef4178 -
Utils/Dataflow/config/009.cfg.example was updated, but 009's readme
was not. Because of this, using the readme's config for launching the
stage would lead to failure due to missing query parametric values.