pyDKB/test: fix '--list' parameter handling.
Function returns last call's return core, which is not necessarily a
zero code ("true"). But when '--list' is used, it is vital to exit after
`list_case()` call; so usage of `&&` was not a good idea (if there was
any issue in `list_case`, or even a simple `if` statement, it can return
non-zero code and command, attached via `&&`, won't be executed: so, we
wouldn't not exit the script, but run test cases.