Add 2018 sample for stage 009.
As ATLAS' development progresses, the data can change. Some types
of new information, such as `hs06sec` for Oracle / Chicago ES, can
be absent in 2016 sample. Therefore, new 2018 sample are added to
properly illustrate the state of the data, but the 2016 one are
preserved because the system must work correctly regardless of the
data gathering time.

Update the stage's README to describe the samples.

Conflicts:
	Utils/Dataflow/009_oracleConnector/README