Improve query parameters error handling.
- A single missing parameter leads to KeyError. Add interception for it
to display a message that points to the configuration.
- The whole configuration missing leads to (even more puzzling)
ORA-00911. Add a warning when no parameters were configured that also
points towards configuration. Since it is possible to use a query
without parameters, no warning should be displayed in such case. Add
a check to only display the warning when the query contains single '%'
(double percent '%%' means that '%' was escaped, so it's allowed).