Store the whole data about chains.
Storing the whole chain of task ids will simplify the future task of
reconstructing the chains.

Remove chain_id and parent_tid as they will be extracted from
chain_data.

Since chain_data is intended for a specific task (and not for
searching/aggregating), the mapping includes special rules for it to
improve the ES' performance. The size of 512 characters should
accomodate the chains up to ~50 tasks long.