es/query: update output sample for htag-steps diagram data.
Changes are caused by some new tasks that was finished since the last
time the sample was generated. It is 4 new tasks: Evgen, Evgen Merge,
Simul, Reco.

As they finished just recently, there are a lot of additional histogram
points added (previously the last point was '2018-09-17', now it is
something like '2019-02-26'.