es/query: improve histogram data preparation (task-steps-gantt).
ES timestamp unit is ms, not sec -- thus 1 day is not 86400, but
86400000.

Number of tasks in a given day is reduced due to the following.
Let `start_time` be `01-01-2018 21:24:07`,
and `end_time` be `02-01-2018 07:18:10`.
In this situation script with interval of `86400` will produce a value,
that goes to `02-01-2018`, while with interval of `86400000` second
value will be outside the interval defined by start/end time.

I believe that we would still loke to have both points in the chart
data, so it is to be fixed.