API: fix issue with 'unhandled exception' (unknown status code).
If status code is not presented in STATUS_CODES hash, it lead to an
exception that won't me caught.

So we will use 'Unknown' in the status line; it will allow client to get
a proper response, just with uninformative description in HTTP status
line (that is not supposed to be read very often... at least not often
enough to change the response, sending information about this bit of
misconfiguration on the server).