API: set 'rtype' parameter to 'json' by default.
Some methods might want to know which type of the response is expected
(to avoid unnecessary operations). Leaving it empty would allow to set
default type for every method individually, but for now it does not look
as a good idea: better to be sure, that every method will return data in
JSON format, not an image or something.