API/task/hist: add start/end time directly to the ES query.
It still does not mean that we don't need to filter data, taken with
this query (as, for example, if `end_time` > `start` and
`start_time` < `start`, we still need to remove dots that are between
`start_time` and `start`). Yet it allows to get less data (e.g. if we
need histogram for 1 month, we don't need to handle all the data
matching given hashtag (as it could be used for years).

Maybe it can be done right within the query (with scripted field instead
of `start_time`/`end_time`), but not right now.