DF/91: allow stage to use DKB ES storage as second (backup) source.
If `--es-config FILE` parameter is specified, use DKB ES storage
as a backup metadata source in case that in primary source (Rucio)
information was removed.

Information from Rucio is removed together with the dataset, so if Rucio
says that dataset is deleted (or not found), the stage can query
internat DKB storage (ES) to get missed information. This functionality
is mostly required for "archived" data reprocessing (when we change the
ETL process or recover missed records).