DF/69: Fix the delimiter issue in read
Unfortunately, the previous solution for 'read' command doesn't work
with some ASCII control characters as EOB like '\n' or '\0'.

"$(echo -ne $EOBatch)" is empty, so a delimiter can't be use.

For that case it's better to use 'eval'.

The 'eval' statement tells the shell to take eval's arguments
and run them through the command-line processing steps all over again.